Add decorators that converts docinstance to docstring
1. Add wrapper for decorator
2. Add decorator that converts an instance of Docstring (stored in _docinstance)
to a docstring in __doc__
3. Add utility function that finds attributes of the given object that is
defined in the same file as the object.
4. Add decorator that recursively converts the Docstring instances of the given
object and all its attributes found via the utility function

Issues:
1. Since the attributes of a function cannot be declared/defined within a
function definition, the Docinstance must be assigned to the function after the
definition which is inconvenient and bad practice. Therefore, it does not seem
practical to completely replace docstrings with instances of Docstring. We can
dig through the python source code to allow Docinstance's in place of the
docstring (string literals), but again, that is impractical.
   Maybe a docstring is necessary, which means it has to be parsed into a
Docstring instance.
2. The main way to find objects associated with the given object is via
__dict__. However, if the module does not have __all__ or __init__, then
the module cannot be recursively decorated.
   Maybe it'll be better to scan the module via directory rather than
through inspection of the object.