Add functions for decorating objects within a module
The recursive decorator (docstring_recursive) on a module only works for objects
that are accessible by the module without additional import's. However, this
would require a specific __init__ structure. To allow different packaging
structures, functions are added such that module's objects can be wrapped
with different __init__ structures.

1. Add documentation for docstring_recursive
2. Add function that wraps all objects defined within the module in which the
function is called.
3. Add function that wraps all imported modules within the any subdirectory of
the directory from which the function is called.
4. Add tests
5. Add fixme comment