Allow strings as types within DocDescription
If we have access to the class of the object being described by
DocDescription, then we can do some nice things like static type
checking. However, this information will not be available if the
DocDescription is constructed by parsing an existing docstring. Though
the docstring parsing is not yet available, we allow types to be
provided as strings. It will not be feasible to try and find the class
that corresponds to the given string.