Make base class for contents of a docstring
The main motivation behind constructing a base class for all content of
a docstring is to establish an API, which should provide a flexible
framework for incorporating different kinds of content into Docstring
without refactoring.

For example, suppose we create another class that constructs a
portion of the docstring for math equations. Since this class can be
used in the extended summary or the description of a variable, this new
class must be incorporated into the DocSection and DocDescription.
Though this is not a problem for one or two new objects, the code will
become bloated as more features are added.

However adding this flexibility results in easier misuse of the code.
This flexibility is not essential yet, so it is not merged into master.

1. Make DocContent class
2. Make DocSection and DocDescription into children of DocContent
3. Move around files and make appropriate changes