Use tmp_path ficture in pytest to create a temporary file
Tox seems to have trouble with finding the dummy python file in the test
directory (module in the tox's virtual environment does not contain the test
directory). Therefore, test_docstring_modify_import failed. To avoid this, a
temporary python file is created instead of storing it in the test directory.

1. Delete dummy file in test directory.
2. Make temporary python files to test docinstance.utils.docstring_modify_import
functionality.
3. Tweak tests to use temporary python files instaed of the dummy file in the
test directory.