Move function's docinstance assignment to inside function
The global keyword can be usd to assign attributes to a function inside the
function. The test for check recursive wrapping of the current module used to
define and assign the function outside of the test function. Now, it can be
moved inside.

The test's answer was also wrong (wrong indentation), but the order in which
Python 3.6 assigns/declares the objects within the code made sure that the
function outside is declared last. Since the function outside would have a
different indentation than if it were defined inside another function/class, the
indentation was one less than desired.