Consolidate multiple docstring methods into a single method DocContent
1. Remove DocContent.make_numpy_docstring, DocContent.make_google_docstring,
DocContent.make_rst_docstring
2. Add DocContent.make_docstring that calls the correct method for the given
style via name of the method
3. Change name of the methods from make_numpy_docstring to make_docstring_numpy,
from make_numpy_docstring_signature to make_docstring_numpy_signature, from
make_google_docstring to make_docstring_google, and from make_rst_docstring to
make_docstring_rst.

Having multiple abstract methods in DocSection for different styles of docstring
seems to get more complicated as the number of styles increases. If the goal is
to make it easier to create docstrings of new styles, then we can just add more
styles to the subclasses without having to add more functions at the base class.

I'm not 100% sure that the best way is to keep adding methods for each
style (this would result in the class growing quite quickly). Maybe it'd be
better to organize the modules by the styles also i.e. make a subclass for each
child of DocSection specific for the style.