Add class for paragraphs
Before, Docstrings/DocSection could handle strings by treating them specially
when the docstring was being created. Normally, blocks of strings were
separately from one another with two newlines wrapped to the appropriate weight
and indented according to user input. However, it seems that sometimes (e.g. in
References or See Also sections) it is necessary to separate out the string
blocks with different number of newlines. Instead of adding a new parameter, a
new class for the string blocks were created, to be consistent with the rest of
the code.

1. Add DocParagraph class
2. Add tests

Time: 2h