Support string content in DocSection
Turns out removing string support causes a lot of the tests to break, so it
would be better to actually support strings in DocSection (and rest of the
code). In fact, it is more intuitive to be able to use strings anyways.

Internally, DocSection still uses DocParagraph, but upon intialization, strings
are converted to DocParagraph